Setting things up
So far, I have to set up the pieces of what- how I want the code structured so I can make the game. There are three different categories and I've made a base class to be addresses and two derived classes. The two derived classes are there since both the board and the pieces are going to be using addresses. Now that I think about it, I could use another- never mind, this is fine.