Address Constructors
The biggest hurdle as of now is the constructor for the address class. It's needed so we can set the address of a chess piece. Tomorrow, I want to look at the chess board and get to work on generating the list every time, including implementing a change at execution time.