Update address.h
I know I said I'd look at something else, but I want to try solving this problem. The concept of having rank and file as enumerators is great, but I can't figure out how to use them in a constructor. I could always make a setRank and a setFile too. Maybe it'd be easier. But the problem wasn't that they're both together- it's that I can't seem t- what if I tried it without a template?

I also tried to do it in a list, but I think I'm missing syntax or something. I'm not so sure what the problem is.