Looked through it again, almost there
So far I just have one error, but I don't know what it's showing up. The error is in the new address.cpp file which I made in efforts to counter another error I was trying to figure out. As of now, I think the error in line 11 means that it doesn't recognize what is on that line as a constructor or an object. I think this might be like this because I used enum? Maybe?