Templatizing, Enumeration, Objects, and Constructors.
The problem was with templatized classes and functions. I now realize I used them with the intention of a data types, when templates are usually used with classes and their objects. In the previous original, address.h, line 5 was meant to use templates for other data types of classes. What I didn't think of was that those data types aren't... classes. The rank and file enumerations are their own data types but not objects. Lines 13-15 were experimental. Lines 17-22 were going to be the alternative if I couldn't solve my initial problem with templates and enumeration. Lines 30-42 is what I tried to do initially.
The stack overflow post that helped me with this was putting almost exactly what I was doing, in the most simplest terms. But it was essentially the same concept. Here is the link: https://stackoverflow.com/questions/34822378/can-i-use-an-enum-in-a-constructor-for-a-class