Need new getString() & rethink container storage
There's two things that I neglected to think about, the getstring and the getid for the containers and pieces data stored in the addresses. If we refer to address.cpp, if I were to return each lines 11-13, then I would get the value (a,b,c, 1,2,3, Empty, Rook... etc), BUT the intention was to ALSO get the getstring. I need a getString function so it would make things easier than to hard-code it- then again, you can't really hard-code because it's something that needs to be compared in run time.
Next:
The container storage might need to be rethought- unless I can remember how to access them either through private or public access of its class.