Continuing Storage Exploration
The list didn't work out very well because I couldn't think of a way to cycle through the list with a loop. So in response, I am going to use array and pointers. I could also use linked lists, but I'm not really inserting information between objects, so I don't need that one.