Default chess piece generator & Chess piece color ID idea
There's a couple of things that need to be pseudocoded, first: how to generate the default set up of the chess pieces on the addresses. Second, which needs attention so that each player moves their corresponding chess pieces and don't move the other player's piece, there needs to be a way to identify which piece corresponds to which player. I could save this in two places. One can be saved in the chess pieces constructor for each player, where the player color piece that corresponds to them is saved. Then, the other place to store the color of a piece is in the address constructor. Then at the end, to compare if a piece corresponds to a user's, you could compare if the pieces have the same color piece or not. So that one, the user can only move that piece, and two, the piece also needs to have that color saved so that it can confirm that they're both the same color.