Rework
Alright. I rethought the storage and access of the addresses and pieces of the files. I'm thinking of making the address more like the chess board where the location is stored. The pieces where the piece and its color is stored, which will, also, confirm the color of each piece for the players. I go more into depth in the next two pages of the pdf.