Creating & Storing Addresses
Here I am making a rough draft of what I want to do. It's a rough draft to correct a couple of things that I neglected to pay attention to when I tried using implementation. So for right now, I'm going to try using a list. This makes it easier to cycle through the different addresses. Now that I realize it, I could actually have a 2d multimap that has both the address and the pieces and such. But as I discussed in the comments, that would make it more difficult to access ONLY those files. Right now my priority is to make things- generally simplify it and then expand on that to build on top of the simple things to make them complicated later on. So, the most important lines here are the while loop on line 87. Lines 87 through 99 is what I want to use to create and store various addresses.